requirements.txt
django~=5.0.7
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary